Get final filemane

Return the filename from the file columns.

Usage

get_final_filename(df)

Arguments

df

The dataframe to be copied. Must have columns locationID, season, roll.

Value

The filename for this file in the format locationID_Sseason_Rroll.csv

It there are several values in locationID, season or roll, they are separated by a dash in the filename: locationID1-locationID2...

Examples

zooniverse_std <- standardize_snapshot_df(df = zooniverse, standard = standard)
#> Initial file: 24 columns, 100 rows.
#> Standardizing columns
#> Standardizing dates/times
#> Fill capture info
#> Cleaning location/camera, species and columns values
#> Final file: 27 columns, 100 rows. Here is a sneak peek:
#> locationID	cameraID	season	roll	eventID	snapshotName	eventDate	eventTime
#> APN	APN_13U	1	1	APN_13U#1#50	kudu	2017-07-25	02:15:22
#> APN	APN_6U	1	1	APN_6U#1#40	steenbok	2017-07-17	08:56:18
#> APN	APN_6U	1	2	APN_6U#2#5	zebraburchells	2017-08-07	04:49:29
#> APN	APN_DW	1	2	APN_DW#2#94	impala	2017-07-06	18:16:19
#> APN	APN_DW	1	1	APN_DW#1#210	zebraburchells	2017-07-21	18:09:16
get_final_filename(zooniverse_std)
#> [1] "APN_S1_R1-2.csv"