Write log message

Writes a log message. If a logger is provided, writes to that logger; if it is NA, displays a message.

Usage

write_log_message(
  message,
  logger = NA,
  level = c("info", "warn", "error", "debug")
)

Arguments

message

The message to display/write to the logger

logger

Logger to write to (log4r object of class logger) (defaults to NA)

level

Logging level: either info, warn, debug or error.

Value

Will either write the message specified in message to the logger or do nothing if logger is NA.

Details

if logger is not NA, writes a message to the logger file with the specified level.

else, doesn't do anything.

Examples

logger <- create_logger(tempfile()) # Create a logger
#> Create logger /tmp/Rtmp3MYUYp/file6d2b7d1f58d0
write_log_message("Test", level = "debug", logger = logger)
write_log_message("Test", level = "info", logger = logger)
write_log_message("Test", level = "warn", logger = logger) 
write_log_message("Test", level = "error", logger = logger)